renaming baseEntity.ts adding missing fields to IAgency.ts
