+++ express-api/src/controllers/IBaseEntity.ts
