+++ express-api/src/controllers/admin/accessRequests/IAccessRequest.ts
@@ -2,7 +2,7 @@ import { UUID } from 'crypto';
 import { IAgency } from '@/controllers/admin/agencies/IAgency';
 import { IRole } from '@/controllers/admin/roles/IRole';
 import { IUser } from '@/controllers/admin/users/IUser';
-import { IBaseEntity } from '@/controllers/baseEntity';
+import { IBaseEntity } from '@/controllers/IBaseEntity';
 
 export interface IAccessRequest extends IBaseEntity {
   id: UUID;
