+++ express-api/src/controllers/admin/claims/IClaim.ts
@@ -1,5 +1,5 @@
 import { UUID } from 'crypto';
-import { IBaseEntity } from '@/controllers/baseEntity';
+import { IBaseEntity } from '@/controllers/IBaseEntity';
 
 export interface IClaim extends IBaseEntity {
   id?: UUID;
