+++ express-api/tests/integration/admin/accessRequests/accessRequests.test.ts
@@ -47,6 +47,9 @@ const expectedList: IAccessRequestList = {
           code: 'string',
           parentId: '3fa85f64-5717-4562-b3fc-2c963f66afa6',
           description: 'string',
+          email: 'test@test.com',
+          sendEmail: true,
+          addreessTo: 'test',
         },
       ],
       roles: [
